+++ index.html
@@ -1,60 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="styles/style.css">
-    <script src="https://kit.fontawesome.com/1e12727c1a.js" crossorigin="anonymous"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <title>TheRipper93</title>
-</head>
-<header class="main-nav">
-        <a class="logo" href="index.html">
-            <img src="assets/logo.jpg" alt="logo">
-            <div class="brand-name">TheRipper93</div>
-        </a>
-        <nav>
-            <ul class="ext-links">
-                <li><a class="filter"><i class="fas fa-filter"></i> Filter</a></li>
-                <li><a class="patreon" href="https://www.patreon.com/theripper93" target="_blank" rel="noopener noreferrer"><i class="fab fa-patreon"></i> Patreon</a></li>
-                <li><a class="discord" href="https://discord.gg/F53gBjR97G" target="_blank" rel="noopener noreferrer"><i class="fab fa-discord"></i> Discord</a></li>
-            </ul>
-        </nav>
-    </header>
-<body class="container full-height-container">
-
-    <section class="cards">
-        <div class="card">
-            <h2 class="card-content">Get access to the module collection for just €5/6$</h2>
-            <ol>
-                <li>Access to 19 modules</li>
-                <li>1+ new modules almost every month</li>
-                <li>Support for my modules</li>
-                <li>Vote on the next modules to be released from Early Access</li>
-                <li>Join Early Access for 3D Canvas and early releases</li>
-            </ol>
-            <a class="card-button patreon-btn" href="https://www.patreon.com/theripper93" target="_blank" rel="noopener noreferrer">Learn More</a>
-        </div>
-        <div class="card">     
-            <h2 class="card-content">Join the community, get guides and resources</h2>
-            <ol>
-                <li>Join many other users of my modules</li>
-                <li>Quick access to videos, tutorials and resources</li>
-                <li>News and updates on my modules</li>
-                <li>Updates on releases from partnered creators</li>
-            </ol>
-            <a class="card-button discord-btn" href="https://discord.gg/F53gBjR97G" target="_blank" rel="noopener noreferrer">Discord</a>
-        </div>
-    </section>
-    <section class="main-content">
-        <ul class="mod-list">
-        </ul>
-    </section>
-    <footer>
-
-    </footer>
-    <script src="scripts/main.js"></script>
-</body>
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/src/favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Vite App</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/main.jsx"></script>
+  </body>
 </html>
