--- scripts/main.js
@@ -1,730 +1,0 @@
-const local = {
-    "en-US": {
-        modules:{
-            "Levels": {
-                name: "Levels",
-                icon: "fas fa-layer-group",
-                desc: "Create maps with multiple vertical levels on the same scene with seamless exploration and combat. Included 3D sight and more.",
-                status: "free",
-                media: "https://youtu.be/ELlweNunn4g",
-                link: "https://github.com/theripper93/Levels",
-              },
-              "Animated Walls": {
-                name: "Animated Walls",
-                icon: "fas fa-university",
-                desc: "Animate your walls with rotation and movement animations. Chain together animations with macros and hook macros on door opening.",
-                status: "free",
-                media: "https://streamable.com/wj3ecv",
-                link: "https://github.com/theripper93/animated-walls",
-              },
-              "Argon Combat HUD": {
-                name: "Argon Combat HUD",
-                icon: "fas fa-shield-alt",
-                desc: "Show a Solasta inspired combat hud for easy accessibility of actions\bonus actions etc.. Repelaces a lot of the character sheet functionality during combat so that you don't need to open it. Basically, token action HUD on steroids.",
-                status: "free",
-                media: "https://streamable.com/ujfxfu",
-                link: "https://github.com/theripper93/enhancedcombathud",
-              },
-              "Automated Evocations": {
-                name: "Automated Evocations",
-                icon: "fas fa-magic",
-                desc: "A user interface to manage companions with summoning animations and automated summoning for spells.",
-                status: "free",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/pax0zn/automated_evocations_companion_manager/",
-                link: "https://github.com/theripper93/automated-evocations",
-              },
-              "Better Roofs": {
-                name: "Better Roofs",
-                icon: "fas fa-building",
-                desc: "Improve overhead tiles by adding funtionality to show them through the fog of war and even masking them based on vision to allow 'peeking' into windows/doors.",
-                status: "free",
-                media: "https://youtu.be/ELlweNunn4g",
-                link: "https://github.com/theripper93/Better-Roofs",
-              },
-              "Boss Bar": {
-                name: "Boss Bar",
-                icon: "fas fa-skull-crossbones",
-                desc: "Create a Dark Souls Like boss healthbar on top of the screen or in other locations. Customize texture, color and more.",
-                status: "free",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/oirzdz/boss_bar_add_soulslike_boss_healthbars_to_your/",
-                link: "https://github.com/theripper93/Boss-Bar",
-              },
-              "Changelogs": {
-                name: "Changelogs",
-                icon: "fas fa-book-open",
-                desc: "Allow devs who opt into the system to display important messages to users after breaking updates, the user can chose which level of warning they wish to recive.",
-                status: "free",
-                media: "",
-                link: "https://github.com/theripper93/libChangelogs",
-              },
-              "Combat Booster": {
-                name: "Combat Booster",
-                icon: "fas fa-bolt",
-                desc: "Speed up your combat with Recent actions, Turn Marker and other smaller tweaks like Pan to combatant, body pile, mark defeated and more. (note that the recent action feature is only dnd5e compatible)",
-                status: "free",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/ot0ek0/combat_booster_speed_up_your_combat/",
-                link: "https://github.com/theripper93/Combat-Booster",
-              },
-              "Dig Down": {
-                name: "Dig Down - Advanced Search",
-                icon: "fas fa-dungeon",
-                desc: "Search inside folders in the file picker, use special search terms to filter your sidebars and many other advanced search filters. Use the excavate functionality to automatically find images for tokens.",
-                status: "free",
-                media: "",
-                link: "https://github.com/theripper93/fuzzy-foundry",
-              },
-              "Go To Player": {
-                name: "Go To Player",
-                icon: "fas fa-arrow-alt-circle-up",
-                desc: "A simple module to get the GM to a specific player (go to it's viewed scene and pan the camera to their token). Simply click the new button in the player list context menu!",
-                status: "free",
-                media: "",
-                link: "https://github.com/theripper93/go-to-player",
-              },
-              "Hover Distance": {
-                name: "Hover Distance",
-                icon: "fas fa-arrows-alt-h",
-                desc: "Show distance to a token on hover without the need to use the ruler. Optionally supports inclusion of vertical distance",
-                status: "paid",
-                media: "https://streamable.com/2z5ijl",
-                link: "",
-              },
-              "Hurry Up": {
-                name: "Hurry Up",
-                icon: "fas fa-hourglass-half",
-                desc: "A simple timer to hurry up your players in combat! Tired of those pesky players taking forever to then just swing a longsword after a 10 minute wait? Well, i have the solution for you!",
-                status: "free",
-                media: "",
-                link: "https://github.com/theripper93/hurry-up",
-              },
-              "IP Shield": {
-                name: "IP Shield",
-                icon: "fas fa-shield-alt",
-                desc: "Simple module to hide the ip fields in the invite links dialog.",
-                status: "free",
-                media: "",
-                link: "https://github.com/theripper93/ip-protect",
-              },
-              "Levels - Automatic Cover Calculator": {
-                name: "Levels - Automatic Cover Calculator",
-                icon: "fas fa-crosshairs",
-                desc: "Calculates a precise cover % that can be used on any system do display a custom message above a token (eg, if only 30% visible display the obscured text). Include options to consider defeated enemies as cover but at a reduced height and more. Preconfigured for dnd5e but works with all systems (Active Effects dnd5e only). Also adds the ability to duck!",
-                status: "paid",
-                media: "https://streamable.com/0hr591",
-                link: "",
-              },
-              "Levels - Volumetric Templates": {
-                name: "Levels - Volumetric Templates",
-                icon: "fas fa-ruler-combined",
-                desc: "Compute Templates in 3D with autotargeting. Automatic mode to dynamically switch between 2d and 3d depending on the information provided. Works with all systems, MIDI QOL compatible.",
-                status: "paid",
-                media: "https://streamable.com/xxt7qf",
-                link: "",
-              },
-              "Patrol": {
-                name: "Patrol",
-                icon: "fas fa-walking",
-                desc: "Automatic or Fixed Patrol Routes for NPCs with the ability to spot player characters. Have your npc wander around to make your city map breath life, have guards on patrol, this module has you covered! (Works very well in conjunction with Auto-Rotate).",
-                status: "free",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/noai9v/patrol_update_the_chase_is_on_free/",
-                link: "https://github.com/theripper93/Patrol",
-              },
-              "RPG Damage Numbers": {
-                name: "RPG Damage Numbers",
-                icon: "fas fa-dice-d20",
-                desc: "Show cascading damage numbers in the style of classic RPGs (for modern reference imagine Borderlands damage). Compatible only with dnd5e, requires midi-qol",
-                status: "paid",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/p1c2jq/rpg_damage_numbers_add_some_style_to_your_combat/",
-                link: "",
-              },
-              "Smart Target": {
-                name: "Smart Target",
-                icon: "fas fa-bullseye",
-                desc: "Target with Alt+Click, For players, target by just left clicking a non-owned token, Show portraits instead of colored pips to indicate targets, positioning/offset and size of the icons can be configured in the module settings, Customize color/shape of the targeting reticule.",
-                status: "free",
-                media: "",
-                link: "https://github.com/theripper93/Smart-Target",
-              },
-              "Splatter": {
-                name: "Splatter",
-                icon: "fas fa-tint",
-                desc: "Add blood and gore to your games, Give that extra touch to your game. Featuring splats, bloodtrails, custom colors and more. Compatible with most game systems (minor configuration required).",
-                status: "free",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/onpf3k/splatter_blood_up_your_games_free_based_on/",
-                link: "https://github.com/theripper93/Splatter",
-              },
-              "Stochastic, Fantastic! - Random Encounter Generator": {
-                name: "Stochastic, Fantastic! - Random Encounter Generator",
-                icon: "fas fa-dice-d20",
-                desc: "A random encounter generator with automatic creature importing (from your own compendiums) and spwaning plus loot generator (with automatic LootSheet creation). dnd5e only.",
-                status: "free",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/p6ydy8/stochastic_fantastic_dnd5e_random_encounter/",
-                link: "https://github.com/theripper93/dnd-randomizer",
-              },
-              "Tile Sort": {
-                name: "Tile Sort",
-                icon: "fas fa-sort-alpha-down",
-                desc: "An easy to use tool to sort the z-index of tiles, includes also a toggle to hide one or more tiles (so you can work on tiles under them) and highlight to tiles on hover (so you can see a tile if it got lost under something). Avaieble for free in the discord channel only.",
-                status: "free",
-                media: "https://streamable.com/6f9ied",
-                link: "",
-              },
-              "Token Z": {
-                name: "Token Z",
-                icon: "fas fa-sort-amount-up",
-                desc: "Easly manage the Z-Index of tokens. The automatic sorting will keep larger tokens on the back and smaller on the front.",
-                status: "free",
-                media: "",
-                link: "https://github.com/theripper93/token-z",
-              },
-              "Weather blocker": {
-                name: "Weather blocker",
-                icon: "fas fa-cloud-sun",
-                desc: "A simple module to block weather effects on a map using drawings. Includes Better Roofs/Levels integration. Works with all systems.",
-                status: "free",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/npixnk/free_weather_blocker_block_weather_through/",
-                link: "https://github.com/theripper93/Weather-Blocker",
-              },
-              "Taskbar": {
-                name: "Taskbar",
-                icon: "fas fa-tasks",
-                desc: "A fully functional Windows style taskbar for foundry, manage your stuff, the Easy way! Include Start style menu and utility icons.",
-                status: "paid",
-                media:
-                  "https://www.reddit.com/r/FoundryVTT/comments/pi0d9i/taskbar_a_windows_style_taskbar_for_foundry/",
-                link: "",
-              },
-              "Quickdraw": {
-                name: "Quickdraw",
-                icon: "fas fa-pencil-alt",
-                desc: "A rapid drawing tool to make quick drawings in-game by holding a hotkey (default q) with live preview for all clients. No more going to the drawing tools, and then deleting stuff! After a set timeout they will automatically fade out.",
-                status: "paid",
-                media: "https://streamable.com/gq5rkn",
-                link: "",
-              },
-              "Config Presets": {
-                name: "Config Presets",
-                icon: "fas fa-boxes",
-                desc: "Save presets for lights/walls/drawings and more, keeps relevant module information and you can apply the presets to existing objects. The ultimate map making tool!",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/pkyui5/config_presets_create_presets_for/",
-                link: "",
-              },
-              "Lightswitch": {
-                name: "Lightswitch",
-                icon: "fas fa-lightbulb",
-                desc: "Simply enable a toggle in the light configuration and now your players will be able to switch on an off the light with a simple click on an icon on the map that is also restricted by sight! No other configuration required.",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/pmu4z0/lightswitch_a_user_frendly_way_to_present/",
-                link: "",
-              },
-              "Journal Drop": {
-                name: "Journal Drop",
-                icon: "fas fa-book-open",
-                desc: "Click a button to place multiple pins of the same journal on the canvas. Each copy will be labeled according to the section of the journal. This will NOT create new journals but use the single existing one ! The splitting is done automagically, in case your journal is not using h1\3 for it's sections you can specify a custom one. Also adds the ability to show journals as tooltips on hover.",
-                status: "paid",
-                media: "https://streamable.com/lstt4w",
-                link: "",
-              },
-              "Filepicker+": {
-                name: "Filepicker+",
-                icon: "fas fa-file",
-                desc: "Expand the default Filepicker to add favourite folders, image and video preview tooltips, audio tooltips and a new layout",
-                status: "paid",
-                media: "https://streamable.com/ql2x55",
-                link: "",
-              },
-              "Progress Tracker": {
-                name: "Progress Tracker",
-                icon: "fas fa-sliders-h",
-                desc: "Create bars with custom background, sliders and waypoints to track many things like Travel, Chases and more. All changes are synced to players!",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/pp3iy2/progress_tracker_create_custom_sliders_to_track/",
-                link: "",
-              },
-              "Situational Shortcuts": {
-                name: "Situational Shortcuts",
-                icon: "fas fa-link",
-                desc: "Whatever you use is, sometimes applying active effects slows down the game. This modules adds quick custom button near the 'Situational Bonus' field. This is perfect for low atutomation users and for those things that are to annoying to automate with active effects! (currently dnd5e only - more systems coming on request)",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/ppoqll/situational_shortcuts_create_quick_buttons_for/",
-                link: "",
-              },
-              "Quick Doors": {
-                name: "Quick Doors",
-                icon: "fas fa-door-open",
-                desc: "Very simple, enable the tool, now clicking near a wall will create a door (on the wall itself) while splitting the original wall appropriatelly, you can also snap to grid to auto allign everything nicely. Includes hotkeys to swap between grid/nogrid/secret/window",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/pqujmq/quick_doors_create_doorswindows_with_1_click/",
-                link: "",
-              },
-              "Image Context": {
-                name: "Image Context",
-                icon: "fas fa-image",
-                desc: "Add a context menu when you right click on any image in foundry. From this menu, show a popout of the image or send it to chat with the ability to wisper it only to some players and more.",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/pzj3dk/image_context_a_convenient_way_to_show_images_to/",
-                link: "",
-              },
-              "Inactive Tokens": {
-                name: "Inactive Tokens",
-                icon: "fas fa-user-slash",
-                desc: "Set token as inactive to make them non interactable, effectively making them indistinguishable from tiles by players. Also, exclude inactive tokens from drag selection and optionally automatically set a token as inactive when defeated.",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/q1h0kv/inactive_tokens_lock_down_tokens_easly_and_make/",
-                link: "",
-              },
-              "3D Canvas": {
-                name: "3D Canvas",
-                icon: "fas fa-cube",
-                desc: "Turn your maps into true 3D, load in 3d models, and play your battles in the glorious third dimension. Also support lighting and animated models. If you don't have 3d models don't worry! 3D canvas will turn your 2d tokens into Stand up figures in the 3d mode.",
-                status: "paidea",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/q2k0en/levels_3d_preview_create_a_3d_view_of_you_levels/",
-                link: "",
-              },
-              "Visual Novel Choices": {
-                name: "Visual Novel Choices",
-                icon: "fas fa-book-reader",
-                desc: "Create a visual novel style choice menu with a list of choices and a title. Make democratic or non votes, customize colors and background image and trigger scene changes or macros when a choice is resolved.",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/q3bqzv/visual_novel_choices_create_vn_style_multiple/",
-                link: "",
-              },
-              "Maxwell's Manual of Malicious Maladies":{
-                name: "Maxwell's Manual of Malicious Maladies",
-                icon: "fas fa-viruses",
-                desc: "A module for all your lingering injuries needs (dnd5e).208 lingering injuries all with different severities, durations, and consequences, based on the 13 different damage types. Includes application automation, durations and item creation to keep track of injuries!. Available only on the discord channel! (for free)",
-                status: "free",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/q7gc5h/free_maxwells_manual_of_malicious_maladies_a/",
-                link: "",
-              },
-              "Socket Macros": {
-                name: "Socket Macros",
-                icon: "fas fa-plug",
-                desc: "A very simple module to execute macros on all clients - simply check the checkbox in the macro configuration and the macro will be fired on all clients!",
-                status: "paid",
-                media: "https://www.reddit.com/r/FoundryVTT/comments/q9bwbj/socket_macros_run_macros_on_all_clients/",
-                link: "",
-              },
-              "Token Flip": {
-                name: "Token Flip",
-                icon: "fas fa-exchange-alt",
-                desc: "Assign multiple token images and actors to the same token and flip between them with ease and a nice animation!",
-                status: "paid",
-                media: "https://youtu.be/W-2EVLscFLM",
-                link: "https://youtu.be/W-2EVLscFLM",
-              },
-              "Token Notes": {
-                name: "Token Notes",
-                icon: "fas fa-sticky-note",
-                desc: "With this simple yet effective module you can take quick notes on tokens. The module is system agnostic and excels at tracking effects durations during combat, especially if you are not using automation!",
-                status: "paidea",
-                media: "https://youtu.be/bb6IS7kt490",
-                link: "",
-              },
-              
-        },
-        free : "Free",
-        paid : "Patreon Only",
-        paidea : "Patreon Only (EA)",
-    },
-
-    "ja": {
-      modules:{
-          "Levels": {
-              name: "Levels（階層モッド）",
-              icon: "fas fa-layer-group",
-              desc: "一つのシーンを複数の階層に分割しシーンを切り替えること無くそれぞれの階層に応じたマップを表示できます。ロード時間を挟まず探索や戦闘も可能です。3Dの視点管理を始めとする様々な便利な機能が入っています。",
-              status: "free",
-              media: "https://youtu.be/ELlweNunn4g",
-              link: "https://github.com/theripper93/Levels",
-            },
-            "Animated Walls": {
-              name: "Animated Walls（アニメーション付きウォール）",
-              icon: "fas fa-university",
-              desc: "普段動きの無いウォールにアニメーションや移動を追加することができます。扉の開け締めにマクロを紐付けることも可能になります。",
-              status: "free",
-              media: "https://streamable.com/wj3ecv",
-              link: "https://github.com/theripper93/animated-walls",
-            },
-            "Argon Combat HUD": {
-              name: "Argon Combat HUD（戦闘画面）",
-              icon: "fas fa-shield-alt",
-              desc: "（翻訳済み）（D&D5版用）キャラクターが使用できるアクション、ボーナスアクション、技能などの自分のキャラクターができることをわかりやすいUIで表示するためのモッドです。これがあれば戦闘中にキャラクターシートを開く必要はなくなります。",
-              status: "free",
-              media: "https://streamable.com/ujfxfu",
-              link: "https://github.com/theripper93/enhancedcombathud",
-            },
-            "Automated Evocations": {
-              name: "Automated Evocations（召喚モッド）",
-              icon: "fas fa-magic",
-              desc: "アクターに別のアクターを紐付けて、アニメーションと共に場に出すことができます。召喚魔法や、他の仲間を召喚するキャラに使えます。（翻訳済み）",
-              status: "free",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/pax0zn/automated_evocations_companion_manager/",
-              link: "https://github.com/theripper93/automated-evocations",
-            },
-            "Better Roofs": {
-              name: "Better Roofs（前景拡張モッド）",
-              icon: "fas fa-building",
-              desc: "前景として設定されたタイルに追加の機能をもたせて、マスクとして除去させたり、窓から覗いたときに覗いたぶんだけを見せるなどの演出が再現できます。",
-              status: "free",
-              media: "https://youtu.be/ELlweNunn4g",
-              link: "https://github.com/theripper93/Better-Roofs",
-            },
-            "Boss Bar": {
-              name: "Boss Bar（ボスHPバー）",
-              icon: "fas fa-skull-crossbones",
-              desc: "ダーク・ソウルシリーズのようなボスのHPバーを画面の上部に表示します。位置や色、大きさ等すべてカスタマイズできます。",
-              status: "free",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/oirzdz/boss_bar_add_soulslike_boss_healthbars_to_your/",
-              link: "https://github.com/theripper93/Boss-Bar",
-            },
-            "Changelogs": {
-              name: "Changelogs（更新ログ）",
-              icon: "fas fa-book-open",
-              desc: "（モッド開発者用）自分のモッドを更新したときにそのモッドを使っているすべてのユーザに大事なお知らせを送ることを可能とします。",
-              status: "free",
-              media: "",
-              link: "https://github.com/theripper93/libChangelogs",
-            },
-            "Combat Booster": {
-              name: "Combat Booster（戦闘拡張モッド）",
-              icon: "fas fa-bolt",
-              desc: "（D&D5版用）ターンを行っているコマにマーカーを表示、最近行ったアクションを表示、ターン開始時にコマにカメラを注目する、自動死亡マーク付与などの便利な機能が満載のモッドです。D&D5版の戦闘の悩みはこれ一つで解決します！",
-              status: "free",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/ot0ek0/combat_booster_speed_up_your_combat/",
-              link: "https://github.com/theripper93/Combat-Booster",
-            },
-            "Dig Down": {
-              name: "Dig Down - Advanced Search（高度な検索）",
-              icon: "fas fa-dungeon",
-              desc: "ファイルを選択するときにフォルダを一つ一つ開くのめんどくさくありませんか？<br>このModを使えばフォルダの中も検索をしてくれます。高度な検索設定でコマに自動的に画像を付与することもできます。",
-              status: "free",
-              media: "",
-              link: "https://github.com/theripper93/fuzzy-foundry",
-            },
-            "Go To Player": {
-              name: "Go To Player（PL注目）",
-              icon: "fas fa-arrow-alt-circle-up",
-              desc: "GMをPLのコマのあるシーンへ移動させてカメラを注目させます。この機能はプレイヤーリストの隣に出るボタンで使用できます",
-              status: "free",
-              media: "",
-              link: "https://github.com/theripper93/go-to-player",
-            },
-            "Hover Distance": {
-              name: "Hover Distance（定規を常時表示）",
-              icon: "fas fa-arrows-alt-h",
-              desc: "コマを動かす時、自動的に定規を出現させます。斜め移動の距離も設定できます。",
-              status: "paid",
-              media: "https://streamable.com/2z5ijl",
-              link: "",
-            },
-            "Hurry Up": {
-              name: "Hurry Up（40秒で支度しな）",
-              icon: "fas fa-hourglass-half",
-              desc: "戦闘中にタイマーをおいて、プレイヤーが早く動くように促すモッドです。",
-              status: "free",
-              media: "",
-              link: "https://github.com/theripper93/hurry-up",
-            },
-            "IP Shield": {
-              name: "IP Shield　（アドレス非表示）",
-              icon: "fas fa-shield-alt",
-              desc: "自分の画面に出てくるグローバルIPを非表示にするモッドです。",
-              status: "free",
-              media: "",
-              link: "https://github.com/theripper93/ip-protect",
-            },
-            "Levels - Automatic Cover Calculator": {
-              name: "Levels - Automatic Cover Calculator（遮蔽自動計算）",
-              icon: "fas fa-crosshairs",
-              desc: "自分のコマから対象のコマへ見えている割合（%）に応じて特別なメッセージを表示するように設定できるモッドです。（例えば30%見えていれば「遮蔽」というメッセージを表示する）<br>死亡した敵も遮蔽として扱えるようにできます。D&D5版用のデフォルト設定を内包しています。<br>しゃがむこともできます！",
-              status: "paid",
-              media: "https://streamable.com/0hr591",
-              link: "",
-            },
-            "Levels - Volumetric Templates": {
-              name: "Levels - Volumetric Templates（3Dテンプレート）",
-              icon: "fas fa-ruler-combined",
-              desc: "テンプレートを3Dに計算し、自動ターゲットも可能にします。与えられた情報に応じて自動的に2D・3Dターゲットを切り替えます。すべてのシステムで利用できます。Midi-qolとの連携も抜群です。",
-              status: "paid",
-              media: "https://streamable.com/xxt7qf",
-              link: "",
-            },
-            "Patrol": {
-              name: "Patrol（パトロール）",
-              icon: "fas fa-walking",
-              desc: "コマにパトロールのルートを設定し、自動的に移動したり、PLを見つけたら近寄って移動するなどの設定ができます。<br>衛兵がパトロールする町を作ってPLにひと味違う遊び方をさせてみませんか？（コマの向きを管理するAuto-rotateモッドと一緒に使うのをお勧めします）",
-              status: "free",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/noai9v/patrol_update_the_chase_is_on_free/",
-              link: "https://github.com/theripper93/Patrol",
-            },
-            "RPG Damage Numbers": {
-              name: "RPG Damage Numbers（ダメージ表示）",
-              icon: "fas fa-dice-d20",
-              desc: "（D&D5版用）Midi-Qol必須のこのモッドは、ダメージを受けたコマの上に数字を表示します（RPGで良くある表示）。",
-              status: "paid",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/p1c2jq/rpg_damage_numbers_add_some_style_to_your_combat/",
-              link: "",
-            },
-            "Smart Target": {
-              name: "Smart Target（スマート・ターゲット）",
-              icon: "fas fa-bullseye",
-              desc: "Alt+クリックでターゲットを可能とし、ターゲットしたコマのアイコンを表示させます。大きさやターゲットの方法等はすべて設定できます。ターゲットマーカーの形も設定可能です。（翻訳済み）",
-              status: "free",
-              media: "",
-              link: "https://github.com/theripper93/Smart-Target",
-            },
-            "Splatter": {
-              name: "Splatter（血糊モッド）",
-              icon: "fas fa-tint",
-              desc: "普段の卓に血糊を追加してみませんか？<br>ダメージを受けたときの血しぶき、移動したときの血痕、色の設定尚d色々と設定できます。すべてのシステムで使用できます（事前設定が必要です）",
-              status: "free",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/onpf3k/splatter_blood_up_your_games_free_based_on/",
-              link: "https://github.com/theripper93/Splatter",
-            },
-            "Stochastic, Fantastic! - Random Encounter Generator": {
-              name: "Stochastic, Fantastic! - Random Encounter Generator（ランダムエンカウント生成）",
-              icon: "fas fa-dice-d20",
-              desc: "（D&D5版用）辞典に登録したNPCやモンスターをエンカウントという一つのセットにまとめて保存できます。まとめて場にだしたり、所持している財宝の自動生成もできます。",
-              status: "free",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/p6ydy8/stochastic_fantastic_dnd5e_random_encounter/",
-              link: "https://github.com/theripper93/dnd-randomizer",
-            },
-            "Tile Sort": {
-              name: "Tile Sort（タイル並び替え）",
-              icon: "fas fa-sort-alpha-down",
-              desc: "タイルのz-軸を設定できるようにするモッドです。特定の軸以下の・以上のタイルを自動的に隠したり、タイルを光らせて選択しやすくできます。",
-              status: "paid",
-              media: "https://streamable.com/6f9ied",
-              link: "",
-            },
-            "Token Z": {
-              name: "Token Z（コマ並び替え）",
-              icon: "fas fa-sort-amount-up",
-              desc: "コマのz-軸を設定できるようにするモッドです。設定していなくても小型のコマは大型のコマの上に自動的に表示されます。",
-              status: "free",
-              media: "",
-              link: "https://github.com/theripper93/token-z",
-            },
-            "Weather blocker": {
-              name: "Weather blocker（天候ブロック）",
-              icon: "fas fa-cloud-sun",
-              desc: "（前景拡張モッド・階層モッドと兼用可能）図形をおいた場所に天候エフェクトをブロックするように設定できます。",
-              status: "free",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/npixnk/free_weather_blocker_block_weather_through/",
-              link: "https://github.com/theripper93/Weather-Blocker",
-            },
-            "Taskbar": {
-              name: "Taskbar（タスクバー）",
-              icon: "fas fa-tasks",
-              desc: "Windowsのようなタスクバーを追加するモッドです。スタートボタンや便利なアイコンを各種追加します。",
-              status: "paid",
-              media:
-                "https://www.reddit.com/r/FoundryVTT/comments/pi0d9i/taskbar_a_windows_style_taskbar_for_foundry/",
-              link: "",
-            },
-            "Quickdraw": {
-              name: "Quickdraw（早描き）",
-              icon: "fas fa-pencil-alt",
-              desc: "ショートカットキーを押している間（デフォルト：Qキー）、接続している全員に見える描画を行う事ができます。描いたものは一定時間の後自動的に消されますので、描いた物を後で消す面倒くさい処理はもういりません！",
-              status: "paid",
-              media: "https://streamable.com/gq5rkn",
-              link: "",
-            },
-            "Config Presets": {
-              name: "Config Presets（まるごと保存）",
-              icon: "fas fa-boxes",
-              desc: "照明・壁・図形などをまるごと保存して、必要なときに複製できます。他のModの設定も保存されるので、一度設定したものを再び設定する必要はありません。",
-              status: "paidea",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/pkyui5/config_presets_create_presets_for/",
-              link: "",
-            },
-            "Lightswitch": {
-              name: "Lightswitch（照明スイッチ）",
-              icon: "fas fa-lightbulb",
-              desc: "照明にオン・オフのスイッチを追加し、扉と同じ要領で明かりを消したりつけたりできます。PLにも扱えます！",
-              status: "paid",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/pmu4z0/lightswitch_a_user_frendly_way_to_present/",
-              link: "",
-            },
-            "Journal Drop": {
-              name: "Journal Drop（ハンドアウト分割）",
-              icon: "fas fa-book-open",
-              desc: "同じハンドアウトの別セクションのコピーを同じシーンに複製します。場に出たものは新たなハンドアウトではなく、すでに存在しているハンドアウトから個別のセクションを使っているので、散らかることもありません。<br>注意：これが正しく機能するには見出しを設定してあげる必要があります。見出しを使っていない場合は独自のセクションを指定できます。",
-              status: "paidea",
-              media: "https://streamable.com/lstt4w",
-              link: "",
-            },
-            "Filepicker+": {
-              name: "Filepicker+（ファイル選択＋）",
-              icon: "fas fa-file",
-              desc: "通常のファイル選択にお気に入りフォルダ、イメージ・動画のプレビュー、見やすいUIを追加するモッドです。",
-              status: "paid",
-              media: "https://streamable.com/ql2x55",
-              link: "",
-            },
-            "Progress Tracker": {
-              name: "Progress Tracker（進行トラッカー）",
-              icon: "fas fa-sliders-h",
-              desc: "チェイス・移動・目標への進行状況を記録するためのバーを表示します。",
-              status: "paidea",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/pp3iy2/progress_tracker_create_custom_sliders_to_track/",
-              link: "",
-            },
-            "Situational Shortcuts": {
-              name: "Situational Shortcuts（状況ボーナス拡張）",
-              icon: "fas fa-link",
-              desc: "（D&D5版用）自動化はしたいけど、完全自動化をするためのややこしい設定をしたくない、数少ない状況ボーナスだけを管理できれば良いという方にお勧めのモッドです。このモッドは事前に設定されたボーナス（+2や+5等）を状況ボーナスの隣にボタンとして表示させ、必要に応じてユーザがクリックして合算してくれます。",
-              status: "paidea",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/ppoqll/situational_shortcuts_create_quick_buttons_for/",
-              link: "",
-            },
-            "Quick Doors": {
-              name: "Quick Doors（クイックドア）",
-              icon: "fas fa-door-open",
-              desc: "ウォールをクリックするだけでドアを作り出せます。元のウォールは自動的に分割されます。マス目あわせも自動的に行われ、シークレットドアや窓へ切り替えるための簡単なショートカットも含まれています。",
-              status: "paid",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/pqujmq/quick_doors_create_doorswindows_with_1_click/",
-              link: "",
-            },
-            "Image Context": {
-              name: "Image Context（画像コンテキスト）",
-              icon: "fas fa-image",
-              desc: "すべての画像に右クリックで出せるメニューを追加し、PLに表示するためのオプションを追加します。チャットに送ったり、特定のPLのみに見せるということもできます。",
-              status: "paidea",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/pzj3dk/image_context_a_convenient_way_to_show_images_to/",
-              link: "",
-            },
-            "Inactive Tokens": {
-              name: "Inactive Tokens（コマロック）",
-              icon: "fas fa-user-slash",
-              desc: "コマを非アクティブとして設定し、動かせないようにするためのモッドです。これを使えばコマが実質タイルになります。敗北したコマを非アクティブにしたり、範囲選択から非アクティブコマを除外してくえｒます。",
-              status: "paidea",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/q1h0kv/inactive_tokens_lock_down_tokens_easly_and_make/",
-              link: "",
-            },
-            "Levels - 3D Preview": {
-              name: "Levels - 3D Preview（階層3Dプレビュー）",
-              icon: "fas fa-cube",
-              desc: "現在のシーンの階層を3Dで表示するための戻っです。3次元にカメラを動かすための操作も内包しています。",
-              status: "paidea",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/q2k0en/levels_3d_preview_create_a_3d_view_of_you_levels/",
-              link: "",
-            },
-            "Visual Novel Choices": {
-              name: "Visual Novel Choices（ビジュアルノベル風選択肢）",
-              icon: "fas fa-book-reader",
-              desc: "ビジュアルノベル風の選択肢を作るためのモッドです。多数決やそれぞれの色の設定も行えます！",
-              status: "paidea",
-              media: "https://www.reddit.com/r/FoundryVTT/comments/q3bqzv/visual_novel_choices_create_vn_style_multiple/",
-              link: "",
-            },
-      },
-      free : "無料",
-      paid : "Patreonのみ",
-      paidea : "Patreonのみ (ベータ)",
-  },
-
-};
-const currentLocal = local[navigator.language] ? local[navigator.language] : local["en-US"];
-const modules = currentLocal.modules;
-
-let filter = 0;
-
-const filterColors = [
-  "rgb(255, 255, 255)",
-  "rgb(255, 98, 71)",
-  "rgb(128, 255, 185)",
-];
-
-const modlist = $(`.mod-list`);
-let sortedMods = [];
-let paidMods = [];
-let freeMods = [];
-for (let [k, v] of Object.entries(modules)) {
-  if (v.status == "paid" || v.status == "paidea") {
-    paidMods.push(k);
-  } else {
-    freeMods.push(k);
-  }
-}
-
-sortedMods.push(...paidMods.sort(), ...freeMods.sort());
-
-for (let mod of sortedMods) {
-  modlist.append(generateLi(modules[mod]));
-}
-
-function generateLi(data) {
-  const status = data.status
-  let statusText = "";
-  if (status == "paid") statusText = currentLocal.paid
-  else if (status == "free") statusText = currentLocal.free
-  else if (status == "paidea") statusText = currentLocal.paidea
-
-  const $li = $(`<li class="module-item" data-status="${data.status}">
-    <i class="${data.icon} mod-icon"></i>
-    <div class="module-info">
-    <a class"module-title-link" href="${
-      data.link
-    }" target="_blank" rel="noopener noreferrer"><h3>${data.name}</h3></a>
-    <p>${data.desc}</p>
-    <div class="info-links">
-        <span>Status: <a href="https://www.patreon.com/theripper93" target="_blank" rel="noopener noreferrer" class="${
-          data.status
-        }">${statusText}</a></span>
-        <a href="${
-          data.media
-        }" target="_blank" rel="noopener noreferrer"><i class="fas fa-play"></i></a>
-    </div>
-    </div>
-    </li>`);
-  return $li;
-}
-
-$(document).on("click", ".filter", function (e) {
-  if (filter == 2) filter = 0;
-  else filter++;
-
-  $(e.currentTarget).css({
-    color: filterColors[filter],
-  });
-  modlist.find(".module-item").each(function (i, el) {
-    switch (filter) {
-      case 0:
-        $(el).slideDown(200);
-        break;
-      case 1:
-        if ($(el).data("status") == "paid") {
-          $(el).slideDown(200);
-        } else {
-          $(el).slideUp(200);
-        }
-        break;
-      case 2:
-        if ($(el).data("status") == "free") {
-          $(el).slideDown(200);
-        } else {
-          $(el).slideUp(200);
-        }
-        break;
-    }
-  });
-});
