+++ src/App.css
@@ -1,0 +1,425 @@
+/** Global Styles **/
+html,
+body {
+  --000: #ffcb77;
+  --100: #ede7d9;
+  --150: #c8c5bf;
+  --200: #a49694;
+  --300: #6f6557;
+  --400: #4b4237;
+  --500: #2d2925;
+  --discord: #7c86ff;
+  --patreon: #ff8970;
+  --free: #74ffb9;
+
+  --herotext: 4rem;
+  --headertext: 2.25rem;
+  --subheadertext: 2rem;
+  --cardtitle: 1.5rem;
+  --buttontext: 1.25rem;
+  --paragraphtext: 1.125rem;
+
+  color: var(--150);
+}
+
+svg {
+  color: var(--150);
+}
+
+input[type='text' i] {
+  background: transparent;
+  color: var(--150);
+
+  border: 0.125rem solid var(--200);
+  width: 78vw;
+  height: 36px;
+}
+
+input:focus-visible {
+  outline-style: none;
+  border-color: var(--000);
+}
+
+/* Scrollbar */
+/* width */
+::-webkit-scrollbar {
+  width: 0.3125rem;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+  background: var(--500);
+}
+
+/* Handle */
+::-webkit-scrollbar-thumb {
+  background: var(--200);
+}
+
+/* Handle on hover */
+::-webkit-scrollbar-thumb:hover {
+  background: var(--000);
+}
+
+/** Text Styles **/
+h1 {
+  font-size: var(--herotext);
+  font-weight: 700;
+}
+
+h2 {
+  font-size: var(--headertext);
+  font-weight: 700;
+}
+
+h3 {
+  font-size: var(--subheadertext);
+  font-weight: 700;
+}
+
+h4 {
+  font-size: var(--cardtitle);
+  font-weight: 700;
+}
+
+h5 {
+  font-size: var(--buttontext);
+  font-weight: 700;
+}
+
+p {
+  font-size: var(--paragraphtext);
+}
+
+/** Navbar **/
+nav {
+  position: fixed;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+
+  overflow: hidden;
+  width: 100%;
+  z-index: 3;
+
+  background: transparent;
+  box-shadow: 0 0.25rem 0.3125rem rgba(24, 20, 20, 0.1);
+  backdrop-filter: blur(0.2938rem);
+  -webkit-backdrop-filter: blur(0.2938rem);
+  font-size: var(--paragraphtext);
+}
+
+nav .website {
+  position: relative;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  align-items: center;
+
+  padding-right: 1rem;
+}
+
+nav .website svg {
+  color: var(--100);
+  padding: 0.125rem;
+}
+
+nav .links {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  align-content: flex-start;
+  justify-content: center;
+  align-items: center;
+
+  padding-left: 1rem;
+
+  color: var(--150);
+  font-size: var(--paragraphtext);
+  font-weight: 500;
+}
+
+nav .links a {
+  color: var(--100);
+}
+
+/** Header and Footer Sections **/
+/* Header */
+section.hero,
+footer {
+  display: flex;
+  position: relative;
+  overflow: hidden;
+
+  min-height: 100vh;
+
+  flex-direction: column;
+  flex-wrap: nowrap;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+  align-self: center;
+
+  background: var(--50);
+}
+
+section.hero h1 {
+  color: var(--150);
+  padding-top: 30vh;
+}
+
+section.hero svg {
+  color: var(--150);
+  margin-top: 30vh;
+  animation: slide 3s ease-in-out infinite;
+}
+
+#video-bg {
+  position: fixed;
+  right: 0;
+  bottom: 0;
+  min-width: 100%;
+  min-height: 100%;
+  z-index: -5;
+}
+
+/* Contact Footer */
+footer {
+  color: var(--150);
+}
+
+footer main {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+}
+
+footer figure {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+  justify-items: center;
+  gap: 1.25rem;
+}
+
+footer main.header-wrapper,
+footer main.header-wrapper h1,
+footer main.header-wrapper h4 {
+  margin-block-start: 0;
+  margin-block-end: 0;
+  margin-inline-start: 0;
+  margin-inline-end: 0;
+}
+
+footer figure .skill-card {
+  width: 11.1875rem;
+  height: 3.25rem;
+
+  display: inline-flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  justify-content: center;
+  text-align: center;
+  align-content: stretch;
+
+  border: 0.1875rem solid var(--200);
+}
+
+footer figure .skill-card h4 {
+  margin-block-start: 0;
+  margin-block-end: 0;
+  margin-inline-start: 0;
+  margin-inline-end: 0;
+  font-weight: bold;
+}
+
+footer main.links {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  align-content: space-between;
+  justify-content: space-between;
+  align-items: center;
+  gap: 30px;
+}
+
+/* Section / Middle Content */
+section {
+  background: var(--500);
+  min-height: 100vh;
+  padding: 80px;
+}
+
+/* Modulelist Section */
+section.module-list {
+  background: var(--500);
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+  flex-wrap: wrap;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.module-list .searchbar {
+  position: relative;
+}
+
+.module-list .searchbar button {
+  position: absolute;
+  right: 1rem;
+  height: 100%;
+  border: 1px solid transparent;
+  background: transparent;
+}
+
+.module-list .searchbar button svg {
+  color: var(--100);
+}
+
+.module-list .searchbar .button-paid svg {
+  color: var(--patreon);
+}
+
+.module-list .searchbar .button-free svg {
+  color: var(--free);
+}
+
+
+.module-grid {
+  position: relative;
+  display: flex;
+  flex-wrap: wrap;
+  padding: 16px;
+  gap: 30px;
+  padding-top: 48px;
+  align-content: space-between;
+  justify-content: center;
+  align-items: center;
+  justify-items: center;
+  flex-direction: row;
+}
+
+/* Divs for the cards */
+.module-card {
+  position: relative;
+  width: 30rem;
+  height: 11rem;
+
+  max-height: 11rem;
+
+  display: flex;
+  flex-direction: row;
+  gap: 12px;
+
+  background: var(--500);
+  border: 2px solid var(--300);
+  filter: drop-shadow(10px 10px 0px var(--300));
+}
+
+.module-card .side-contents {
+  display: flex;
+  position: relative;
+  flex-direction: column;
+  justify-content: center;
+  min-width: 25%;
+  max-width: 25%;
+}
+
+.module-card svg {
+  padding: 1.875rem;
+  align-self: center;
+  color: var(--150);
+}
+
+.module-card .card-contents {
+  position: relative;
+}
+
+.module-card h4 {
+  margin-block-start: 0;
+  margin-block-end: 0;
+  margin-inline-start: 0;
+  margin-inline-end: 0;
+
+  padding-top: 1.25rem;
+}
+
+.module-card article {
+  max-height: 5.5rem;
+  max-width: auto;
+  padding-right: 1rem;
+  overflow: hidden;
+}
+
+.module-card .paidea,
+.module-card .paid,
+.module-card .free {
+  position: absolute;
+  margin-block-end: 0;
+  margin-block-start: 0;
+
+  bottom: 0.5rem;
+  right: 0.5rem;
+  padding: 0;
+  text-align: right;
+  width: 100%;
+  font-size: 1rem;
+  font-weight: 500;
+}
+
+.module-card .paidea {
+  color: var(--000);
+}
+
+.module-card .paid {
+  color: var(--patreon);
+}
+
+.module-card .free {
+  color: var(--free);
+}
+
+/* Call to Action / Subscribe */
+section.call-to-action {
+  display: flex;
+  gap: 5rem;
+  flex-direction: row;
+  flex-wrap: wrap;
+  align-content: flex-start;
+  justify-content: center;
+  align-items: center;
+}
+
+section.call-to-action .patreon,
+section.call-to-action .discord {
+  width: 25%;
+}
+
+section.call-to-action .patreon h3 {
+  color: var(--patreon);
+}
+
+section.call-to-action .discord h3 {
+  color: var(--discord);
+}
+
+/* Animation */
+@keyframes slide {
+  0%,
+  100% {
+    transform: translate(0, 0);
+  }
+
+  50% {
+    transform: translate(0, 0.625rem);
+  }
+}
