+++ src/Components/NavLinks.jsx
@@ -1,0 +1,13 @@
+const NavLinks = () => {
+  return (
+    <div className='links'>
+    <p>Home</p>
+    <a> / </a>
+    <p>Modules</p>
+    <a> / </a>
+    <p>Installation</p>
+    </div>
+  )
+}
+
+export default NavLinks
