+++ src/Components/SkillCard.jsx
@@ -1,0 +1,9 @@
+const SkillCard =  (props) => {
+  return(
+    <div className="skill-card">
+        <h4>{props.name}</h4>
+    </div>
+  )
+}
+
+export default SkillCard
