--- styles/style.css
@@ -1,286 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Poppins:400,300,600,700);
-
-:root{
-    --font-color: #fff;
-}
-
-*, *::before, *::after {
-  box-sizing: border-box;
-}
-
-body{
-    background-color: rgb(46, 46, 61);
-    padding: 0;
-    margin: 0;
-    font-family: 'Poppins', sans-serif;
-    font-size: 1.2rem;
-    color: var(--font-color);
-    min-height: 100vh;
-}
-
-header{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 120px;
-}
-
-.full-height-container{
-    display: flex;
-    flex-direction: column;
-}
-
-.container{
-    width: 80vw;
-    margin: 0 auto;
-    padding: 0 40px;
-}
-
-.logo{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.25rem;
-    text-decoration: none;
-    color: inherit;
-}
-
-.logo img{
-    width: 50px;
-    height: 50px;
-    margin-right: 1rem;
-    border-radius: 50%;
-    border: none;
-}
-
-.ext-links{
-    list-style: none;
-    display: flex;
-    flex-direction: row;
-    font-size: 1.5rem;
-    margin: 0;
-    padding: 0;
-    text-decoration: none;
-}
-
-.ext-links a{
-    text-decoration: none;
-}
-
-.ext-links li{
-    margin-right: 1rem;
-}
-
-.discord{
-    color: rgb(103, 86, 255);
-}
-
-.patreon{
-    color: rgb(248, 92, 64);
-}
-
-.filter{
-    cursor: pointer;
-}
-
-.install{
-    color: rgb(64, 248, 156);
-}
-
-.main-content{
-    flex-grow: 1;
-}
-
-.mod-list{
-    list-style: none;
-    margin: 0;
-    padding: 0;
-}
-
-.module-item{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background-color: rgba(0, 0, 0, 0.26);
-    border-radius: 10px;
-    margin-bottom: 1rem;
-    padding: 0 1rem;
-    padding-bottom: 1rem;
-}
-
-.module-item i{
-    font-size: 5rem;
-    margin-right: 1rem;
-    padding: 1rem;
-}
-
-
-.module-item a{
-    text-decoration: none;
-}
-
-.module-item h3{
-    font-size: 1.5rem;
-    color: var(--font-color);
-}
-
-.free{
-    color: rgb(128, 255, 185);
-}
-
-.paid, .paidea{
-    color: rgb(255, 98, 71);
-}
-
-.info-links{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.info-links i{
-    color: var(--font-color);
-    text-decoration: none;
-    font-size: 1.5rem;
-}
-
-.module-info{
-    flex-grow: 1;
-}
-
-.mod-icon{
-    min-width: 7rem;
-    text-align: center;
-}
-
-
-
-
-
-.main-nav{
-    position: fixed;
-    background-color: rgb(46, 46, 61);
-    width: calc(80vw - 80px);
-    height: 5rem;
-    padding: 0 1rem;
-}
-
-/*CARDS*/
-
-.cards{
-    display: flex;
-    justify-content: space-evenly;
-    padding-top: 6rem;
-    padding-bottom: 3rem;
-    max-height: 100vh;
-}
-
-.card{
-    display: flex;
-
-    align-items: center;
-    flex-direction: column;
-    padding: 2rem;
-    max-width: 40%;
-    color: black;
-    background-color: white;
-    border-radius: 10px;
-    box-shadow: 0 0 5px black;
-}
-
-.card h2{
-    text-align: center;
-    font-size: 1.8vw;
-}
-
-.card ol{
-    list-style: disc; 
-    align-self: center;
-    padding: 1rem;
-    font-size: 1vw;
-}
-
-.card-button{
-    margin-top: auto;
-    text-decoration: none;
-    color: black;
-    font-weight: bold;
-    background: none;
-    border: none;
-    background-color: rgb(248, 92, 64);
-    padding: 1rem 2rem;
-    font-size: 1vw;
-    border-radius: 5px;
-    box-shadow: 0 0 0 0 rgb(248, 92, 64);
-    transition: box-shadow ease-in-out 0.2s, filter ease-in-out 0.2s;
-}
-.card-button.patreon-btn{
-    background-color: rgb(248, 92, 64);
-    box-shadow: 0 0 0 0 rgb(248, 92, 64);
-}
-.card-button.patreon-btn:hover{
-    box-shadow: 0 0 0 5px rgb(248, 92, 64);
-}
-.card-button.discord-btn{
-    background-color: rgb(103, 86, 255);
-    box-shadow: 0 0 0 0 rgb(103, 86, 255);
-}
-.card-button.discord-btn:hover{
-    box-shadow: 0 0 0 5px rgb(103, 86, 255);
-}
-.card-button:hover{
-    cursor: pointer;
-    box-shadow: 0 0 0 5px rgb(248, 92, 64);
-    filter: brightness(1.2);
-}
-
-
-@media screen and (max-width: 768px){
-    .module-item i{
-        font-size: 0.6rem;
-    }
-    .mod-icon{
-        display: none !important;
-    }
-    .ext-links{
-        font-size: 0.8rem !important;
-    }
-    .brand-name{
-        display: none;
-    }
-    body{
-        font-size: 0.6rem;
-    }
-    .container{
-        width: 100vw;
-        margin: 0;
-        padding: 0 10px;
-    }
-    .module-item h3{
-        font-size: 0.7rem;
-    }
-    .cards{
-        flex-direction: column;
-        align-items: center;
-        height: auto;
-        gap: 1rem;
-    }
-    .card{
-        max-width: 80%;
-    }
-    .main-nav{
-        width: 100vw;
-    }
-}
-
-@media screen and (max-width: 1200px){
-    .cards{
-        flex-direction: column;
-        align-items: center;
-        height: auto;
-        gap: 1rem;
-    }
-    .card{
-        max-width: 80%;
-    }
-}
